pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "content-engine"
version = "0.1.0"
description = "Automated short-form viral video pipeline for TikTok, Snapchat Spotlight, and YouTube Shorts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "python-dotenv>=1.0.1",
]

[project.scripts]
content-engine = "content_engine.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
